# docs/swagger.yaml
basePath: /
definitions:
  domain.AbilityRequest:
    properties:
      ability:
        type: string
      target_id:
        type: integer
    type: object
...
  domain.CreateRoomRequest:
    properties:
      type:
        type: string
    type: object
  domain.GameRoom:
    properties:
      code:
        type: string
      created_at:
        type: string
      day_count:
        type: integer
      deleted_at:
        type: string
      host_id:
        type: integer
      id:
        type: integer
      phase:
        type: string
      players:
        items:
          $ref: '#/definitions/domain.User'
        type: array
      results:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
      winner:
        type: string
    type: object
...
  domain.LoginRequest:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
  domain.Profile:
    properties:
      age:
        type: integer
      avatar:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      friends:
        type: integer
      gender:
        type: string
      id:
        type: integer
      level:
        type: integer
      losses:
        type: integer
      medals:
        items:
          type: string
        type: array
      name:
        type: string
      play_time:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      wins:
        type: integer
    type: object
...
  domain.RegisterRequest:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
...
  domain.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      otp:
        type: string
      otp_expires:
        type: string
      password:
        type: string
      phone:
        type: string
      profile:
        $ref: '#/definitions/domain.Profile'
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
      wallet:
        $ref: '#/definitions/domain.Wallet'
    type: object
  domain.VerifyOTPRequest:
    properties:
      age:
        type: integer
      gender:
        type: string
      name:
        type: string
      otp:
        type: string
      phone:
        type: string
    required:
    - otp
    - phone
    type: object
  domain.VoteRequest:
    properties:
      target_id:
        type: integer
    type: object
  domain.Wallet:
    properties:
      coins:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      diamonds:
        type: integer
      id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the Mafia game.
  title: Mafia API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Sends a one-time password to authenticate an existing user.
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request OTP for login
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Sends a one-time password to start a new account registration.
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request OTP for registration
      tags:
      - Auth
  /auth/verify:
    post:
      consumes:
      - application/json
      description: Confirms the OTP sent to the user and returns an access token.
      parameters:
      - description: OTP verification payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify an OTP code and issue a token
      tags:
      - Auth
  /game/rooms:
    get:
      description: Lists available game rooms.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.GameRoom'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List game rooms
      tags:
      - Game
    post:
      consumes:
      - application/json
      description: Creates a new game room for the authenticated user.
      parameters:
      - description: Room payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GameRoom'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a game room
      tags:
      - Game
  /game/rooms/{id}/ability:
    post:
      consumes:
      - application/json
      description: Uses a role ability in an active game.
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ability payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.AbilityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Use an ability
      tags:
      - Game
  /game/rooms/{id}/join:
    post:
      description: Adds the authenticated user to a room by ID.
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Join a game room
      tags:
      - Game
  /game/rooms/{id}/leave:
    post:
      description: Removes the authenticated user from a room by ID.
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Leave a game room
      tags:
      - Game
  /game/rooms/{id}/phase:
    post:
      description: Advances the game phase for a room.
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GameRoom'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Advance game phase
      tags:
      - Game
  /game/rooms/{id}/start:
    post:
      description: Starts the game for a given room ID.
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Start a game
      tags:
      - Game
  /game/rooms/{id}/vote:
    post:
      consumes:
      - application/json
      description: Records a vote in an active game.
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Vote payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.VoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit a vote
      tags:
      - Game
...
swagger: "2.0"